# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Teacher was successfully logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginOutputDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "If the password or login or email is wrong"
          },
          "429": {
            "description": "More than 5 attempts from one IP-address during 10 seconds"
          }
        },
        "summary": "Login teacher",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/register-admin": {
      "post": {
        "operationId": "AuthController_registerAdmin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Admin registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterAdminOutputDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "More than 5 attempts from one IP-address during 10 seconds"
          }
        },
        "summary": "Registrate a new admin",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh-token": {
      "post": {
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Token pair was successfully refreshed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginOutputDto"
                }
              }
            }
          },
          "401": {
            "description": "If the refresh token is wrong or expired"
          },
          "429": {
            "description": "More than 5 attempts from one IP-address during 10 seconds"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate new pair of jwt tokens",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "School API",
    "description": "The School API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "LoginOutputDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "The access token",
            "example": "1234567890"
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "BadRequestErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "The HTTP status code",
            "example": 400
          },
          "path": {
            "type": "string",
            "description": "The path of the request",
            "example": "/api/v1/users"
          },
          "message": {
            "type": "object",
            "description": "The message of the error",
            "example": [
              "Error message 1",
              "Error message 2"
            ]
          }
        },
        "required": [
          "statusCode",
          "path",
          "message"
        ]
      },
      "RegisterAdminDto": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "description": "The login of the admin",
            "example": "admin123"
          },
          "password": {
            "type": "string",
            "description": "The password of the admin",
            "example": "password123"
          },
          "name": {
            "type": "string",
            "description": "The name of the admin",
            "example": "John Doe"
          },
          "secret_key": {
            "type": "string",
            "description": "The secret key for the admin",
            "example": "1234567890"
          }
        },
        "required": [
          "login",
          "password",
          "name",
          "secret_key"
        ]
      },
      "RegisterAdminOutputDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "The message",
            "example": "Admin registered successfully"
          }
        },
        "required": [
          "message"
        ]
      },
...
    }
  }
}